Fix a bug with migrating null regular columns
If the migrator encounters a row containing entirely null values in
the regular columns, it would skip over that row, whereas it should
have inserted the primary key values.
This is due to the logic in `explodeRow` skipping all rows with a
null writetime, which all null-valued columns have.

An example row that is affected by the bug:

    CREATE TABLE test (key int PRIMARY KEY, value INT);
    INSERT INTO TEST (key) VALUES (1);

The new logic allows inserting rows with writetime marked as unset.